User permission API endpoint added
