+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/vertx/MainVerticle.java
@@ -95,6 +95,7 @@ public class MainVerticle extends AbstractVerticle {
     builder.operation("topicConfiguration").handler(_endPoints::handle);
     builder.operation("admin").handler(_endPoints::handle);
     builder.operation("rightsize").handler(_endPoints::handle);
+    builder.operation("permissions").handler(_endPoints::handle);
     builder.rootHandler(StaticHandler
             .create()
             .setCachingEnabled(false)
