+++ cruise-control/src/test/java/com/linkedin/kafka/cruisecontrol/servlet/UserPermissionsManagerTest.java
@@ -7,7 +7,7 @@ package com.linkedin.kafka.cruisecontrol.servlet;
 import com.linkedin.kafka.cruisecontrol.KafkaCruiseControlUnitTestUtils;
 import com.linkedin.kafka.cruisecontrol.config.KafkaCruiseControlConfig;
 import com.linkedin.kafka.cruisecontrol.config.constants.WebServerConfig;
-import org.jetbrains.annotations.NotNull;
+import javax.validation.constraints.NotNull;
 import org.junit.Test;
 import java.util.Arrays;
 import java.util.Collections;
